Add PostgreSQL 9.6.6 as a service.
This will be configured as a service for more than one container.

This setting sets up the admin user as user 'baron'. It does not set an initial password for this user. This should be done afterwards.
I have chosen to expose the database port to maui itself. This will make it easier to tunnel the database through ssh, as well as to make (manual) changes to various databases without having to enter into a container.