Update Trac image and add to docker-compose (#10)
* Update Trac image with manual build of the psycopg2 module.

The alpine base image has python installed both with the /usr and the /usr/local prefix. pip is only installed in the latter one. This approach builds all the additional requirements in the /usr/local prefix. For psycopg2 this means that we need a build environment, plus the postgresql development package

* Add Trac to the services running on maui

* Add certificates for dev.haiku-os.org