Trac backup script and better mounting (#12)
* Mount the Trac with the :z parameter
The SELinux configuration of the host by default does not allow writing on the /var directory. This parameter makes sure that the proper permissions are set up.

* Add a backup script for the Trac instance

This script performs a manual pg_dump. This is done because the pg_dump
version has to match the server version. By doing the dump manually
(and not through the trac-admin hotcopy command) we avoid being
dependent on a specific version.