Gerrit: include the full diff in notification emails
Restores part of the workflow we had before Gerrit for reviewing
directly on the mailing list. The other half is allowing Gerrit to
receive replies to these emails, which needs a mail account somewhere
and configuring the receiveemail.* settings.

Then people could reply on the mailing list, and Gerrit would parse
their replies and convert them to review comments directly, so reviewer
wouldn't need to use the web interface.