Concourse: add catkeys task to haiku-release pipeline (#73)
This task creates the English catalogues for the current checkout, zips them up, and uploads them to a bucket that holds the latest translations.

Currently, these catkeys are built per branch, per architecture. This will incur a slight overhead, as the translation tooling will only use the x86_gcc2h arch, but the size of these files are small so it is acceptable.

Only the latest catkeys build is retained. This gives the advantage of a predictable filename. There also is no need to keep other catkey archives.

Part of:
 * https://github.com/haiku/infrastructure/issues/37
 * https://dev.haiku-os.org/ticket/17530