Deployment for Haiku Format Bot
The template is based on various other templates.

The `haiku-format-bot` is intended to run as a deamon that monitors Gerrit, and
applies `haiku-format` to new patches. The process is designed to run
continuously. There is no user facing frontend (so there is no web server
required). The resource limits have been copied from the `irccat.yml` file.
The bot is stateless, so there are no volumes. The bot should in theory be able
to run with multiple replicas, but there is no point in doing so.

In the early stages, it will be monitored manually to make sure that it keeps
running and to iron out issues.